Add a `queuefilled` event
This event allow to notify the queue is filled of files
and ready to be processed. The use case is to have a semi
automatically way of uploading file. You only want
to do some thing just before uploading it so you can't
use `autoProcessQueue` feature but you don't want to let
the user do it.

Example:

```coffeescript
drop = new Dropzone '#dropzone',
        autoProcessQueue: false
drop.on 'queuefilled', (files) ->
        # do what you want
	drop.processQueue()
```